Migrate to AGP 9 KMP android plugin and replace bignum with Double
Switch all modules from the deprecated `androidLibrary {}` block to the
new `android {}` block required by AGP 9's `com.android.kotlin.multiplatform.library`
plugin. Restore `withAndroidTarget()` alongside the existing `withCompilations`
workaround for issue 442950553 so the IDE model is set up correctly.

The new plugin requires KMP libraries to publish an androidTarget artifact.
`com.ionspin.kotlin:bignum` only publishes jvm/native variants, so replace
its `BigDecimal` usage in the scripting module with stdlib `Double`. Precision
is more than adequate for WSL scripting arithmetic (counters, timers, comparisons).

+++ scripting/src/commonTest/kotlin/wsl/WslOperatorTest.kt
@@ -1,6 +1,5 @@
 package warlockfe.warlock3.scripting.wsl
 
-import com.ionspin.kotlin.bignum.decimal.toBigDecimal
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
@@ -8,7 +7,7 @@ import kotlin.test.assertFalse
 import kotlin.test.assertTrue
 
 class WslOperatorTest {
-    private fun num(value: Int) = WslNumber(value.toBigDecimal())
+    private fun num(value: Int) = WslNumber(value.toDouble())
 
     // --- WslAdditiveOperator ---
